MAJOR CODE REFACTORING
