+++ frontend/src/App.js
@@ -1,14 +1,13 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import "./index.js";
-import Login from "./components/login.js";
-import SignUp from "./components/signup.js";
-
-import TopRated from "./components/top_rated.js";
-import NowPlaying from "./components/now_playing.js";
-import Popular from "./components/popular.js";
-import Upcoming from "./components/upcoming.js";
-import Homepage from "./components/homepage.js";
+import Login from "./views/login.js";
+import SignUp from "./views/signup.js";
+import TopRated from "./views/top_rated.js";
+import NowPlaying from "./views/now_playing.js";
+import Popular from "./views/popular.js";
+import Upcoming from "./views/upcoming.js";
+import Homepage from "./views/homepage.js";
 import PrivateRoute from "./components/privateRoutes.js";
 function App() {
   return (
