+++ frontend/src/misc/moviecard.scss
@@ -4,13 +4,13 @@ $brand-color: #fe4141;
 $dark-color: #1e1b26;
 
 body {
-  background: linear-gradient(
-    rgba(255, 255, 255, 0.95),
-    rgba(30, 27, 38, 0.95)
-  );
-  background-position: center;
-  background-size: cover;
-  background-repeat: repeat;
+  //   background: linear-gradient(
+  //     rgba(255, 255, 255, 0.95),
+  //     rgba(30, 27, 38, 0.95)
+  //   );
+  //   background-position: center;
+  //   background-size: cover;
+  //   background-repeat: repeat;
   overflow-y: auto;
 }
