+++ frontend/src/views/error.js
