+++ frontend/src/views/now_playing.js
