+++ frontend/src/views/top_rated.js
